feat(modal): Implement modal directive.
This is different from the [angular-ui/bootstrap](https://github.com/angular-ui/bootstrap)
implementation of modal in a number of ways:

1) No support for stacked modal windows.
2) Can be activated by setting a model value via the is-active attribute.

It is planned to provide a service for the modal, which will allow it to perform more advanced
tasks.

Additionally, custom modal controllers should be easily doable. These can wait a night, however.

There are no tests as of yet, they will be provided before merging this. However, in a little
test HTML page, it is working wonderfully so far. Wunderbar.