feat(dropdown): add dropdown and dropdown-menu directives
Initial work on some AngularJS ports of the dropdown modules.

As of yet, there is no real test suite, and testing has only been
happening in what I intend to be a future demo site for the
AngularSemantic codebase.

Currently, transition support is experimental, using $animate's
addClass methods exclusively. I'd like to make this a bit more
dynamic, and additionally support jQuery animation functions if
jQuery is detected -- But for now, this is working nicely, I think.