fix(checkbox): don't ignore click event for outer element
There's an issue where wrapping the checkbox element in a <label>
causes problems, because the click event is sent to each child,
causing the events to cancel each other out.

This is a problem, but for the moment I don't have a proper fix,
and this current solution seems to break the unwrapped behaviour.

So, I guess it's necessary to find an alternative solution for this.